+++ app/src/main/java/com/hongstudio/kakaosearchimage/database/DocumentDao.kt
@@ -1,9 +1,10 @@
-package com.hongstudio.kakaosearchimage
+package com.hongstudio.kakaosearchimage.database
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
+import com.hongstudio.kakaosearchimage.model.Document
 
 @Dao
 interface DocumentDao {
