+++ app/src/main/java/com/hongstudio/kakaosearchimage/database/DocumentDatabase.kt
@@ -1,9 +1,10 @@
-package com.hongstudio.kakaosearchimage
+package com.hongstudio.kakaosearchimage.database
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.hongstudio.kakaosearchimage.model.Document
 
 @Database(entities = [Document::class], version = 1)
 abstract class DocumentDatabase : RoomDatabase() {
