+++ app/src/main/res/layout/fragment_favorite.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FavoriteFragment">
+    tools:context=".view.FavoriteFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewImageList"
